PHAL: istep mode: hardware error is not stopping boot process.
In the istep mode an error is logged with hardware error
should stop the boot process and stays in Quiesce state  if the
stop on error  policy is enabled.
Noticed diffrent behaviour in istep mode, like Quiesce calls
phal-create-boottime-guard-indicator.service. This service
starts start_host@0.service, causing normal boot start in the
middle of a Quiesce.

In most cases, this would be fine, but we didn't use firmware
to boot so we did not run "start_host@0.service".

Proposed fix is to remove Wants=start_host@0.service

Tested: forced the checkstop during istep boot.
        system stay's at Quiesce state.

Signed-off-by: Jayanth Othayoth <ojayanth@in.ibm.com>
Change-Id: I4c11648f9162e2c0c861cb1c96edafd6e22323b6