Add jobserver support branch
Ninja has a PR for adding make jobserver support [1] that has been a widely
debated PR for many... many years.  Given that many people have forked to
incorporate this PR, and it claims to solve a problem we have (OOM on gcc
processes) it seems like it would be worthwhile using a well maintained fork
instead of the main project.

This is not a one way door.  If we find that the project goes
unmaintained, doesn't build, or otherwise has problems, we can always go
back to using mainline.

Of the forks that have pulled this in, there are:
The Fuscia project [2]
Their targets seem more specific and less generic, although their
improvements seem more extensive.

Kitware [3]
Maintains a fork of ninja

Docker [4]

[1] https://github.com/ninja-build/ninja/issues/1139
[2] https://fuchsia.googlesource.com/third_party/github.com/ninja-build/ninja/+/refs/heads/main/README.fuchsia
[3] https://github.com/Kitware/ninja
[4] https://github.com/dockbuild/ninja-jobserver

'''
EXTRA_OEMESON_COMPILE:append = " \
--ninja-args='--tokenpool-master=fifo' \
"

PARALLEL_MAKE = "-j 20"
BB_NUMBER_THREADS = "20"
'''

Signed-off-by: Ed Tanous <ed@tanous.net>